refactor: extract RocketChatMessageClient from RocketChatService
